fix(arm32): narrow fpscr to uint32_t when copying to CPUContextARM
After widening MinidumpContextARM::fpscr to uint64_t to match
breakpad's on-disk layout, the in-memory CPUContextARM::vfp_regs.fpscr
is still uint32_t (the ARM FPSCR register is genuinely 32 bits; the
u64 in the minidump stream is padding for alignment, not extra data).
Add an explicit static_cast to silence MSVC's /WX narrowing warning.

+++ snapshot/minidump/minidump_context_converter.cc
@@ -158,7 +158,7 @@ bool MinidumpContextConverter::Initialize(
     context_.arm->lr = src->lr;
     context_.arm->pc = src->pc;
     context_.arm->cpsr = src->cpsr;
-    context_.arm->vfp_regs.fpscr = src->fpscr;
+    context_.arm->vfp_regs.fpscr = static_cast<uint32_t>(src->fpscr);
 
     for (size_t i = 0; i < std::size(src->vfp); i++) {
       context_.arm->vfp_regs.vfp[i] = src->vfp[i];
